video_service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.20
PyJWT